Moved utils dir to root
