+++ utils/copyright_variables.js
